+++ src/python/pants/core/register.py
@@ -10,6 +10,7 @@ from pants.core.goals import (
     check,
     export,
     fmt,
+    generate_lockfiles,
     lint,
     package,
     publish,
@@ -50,6 +51,7 @@ def rules():
         *check.rules(),
         *export.rules(),
         *fmt.rules(),
+        *generate_lockfiles.rules(),
         *lint.rules(),
         *update_build_files.rules(),
         *package.rules(),
